App Split Logic functional
