+++ data/data.js
@@ -1,0 +1,146 @@
+// data.js
+
+export const libraryBooks = [
+  {
+    title: "Moby Dick",
+    author: "Herman Melville",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Moby+Dick" },
+  },
+  {
+    title: "Frankenstein",
+    author: "Mary Shelley",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Frankenstein" },
+  },
+  {
+    title: "Romeo and Juliet",
+    author: "William Shakespeare",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Romeo+and+Juliet",
+    },
+  },
+  {
+    title: "Pride and Prejudice",
+    author: "Jane Austen",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Pride+and+Prejudice",
+    },
+  },
+  {
+    title: "Alice's Adventures in Wonderland",
+    author: "Lewis Carroll",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Alice" },
+  },
+];
+
+
+export const recommendations = [
+  {
+    title: "Little Women",
+    author: "Louisa May Alcott",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Little+Women" },
+  },
+  {
+    title: "Dracula",
+    author: "Bram Stoker",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Dracula" },
+  },
+  {
+    title: "The Strange Case of Dr. Jekyll and Mr. Hyde",
+    author: "Robert Louis Stevenson",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Jekyll+%26+Hyde",
+    },
+  },
+  {
+    title: "The Great Gatsby",
+    author: "F. Scott Fitzgerald",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Gatsby" },
+  },
+  {
+    title: "A Tale of Two Cities",
+    author: "Charles Dickens",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Tale+of+Two+Cities",
+    },
+  },
+  {
+    title: "Adventures of Huckleberry Finn",
+    author: "Mark Twain",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Huck+Finn" },
+  },
+  {
+    title: "Wuthering Heights",
+    author: "Emily Brontë",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Wuthering+Heights",
+    },
+  },
+  {
+    title: "War and Peace",
+    author: "Leo Tolstoy",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=War+and+Peace",
+    },
+  },
+  {
+    title: "Grimms' Fairy Tales",
+    author: "Jacob and Wilhelm Grimm",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Grimms+Tales" },
+  },
+  {
+    title: "The Adventures of Sherlock Holmes",
+    author: "Arthur Conan Doyle",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Sherlock+Holmes",
+    },
+  },
+  {
+    title: "The Works of Edgar Allan Poe — Volume 2",
+    author: "Edgar Allan Poe",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Edgar+Allan+Poe+V2",
+    },
+  },
+  {
+    title: "The Iliad",
+    author: "Homer",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=The+Iliad" },
+  },
+  {
+    title: "The Importance of Being Earnest",
+    author: "Oscar Wilde",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Earnest" },
+  },
+  {
+    title: "The Count of Monte Cristo",
+    author: "Alexandre Dumas",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Monte+Cristo" },
+  },
+  {
+    title: "Anne of Green Gables",
+    author: "L. M. Montgomery",
+    cover: {
+      uri: "https://via.placeholder.com/120x180.png?text=Anne+of+Green+Gables",
+    },
+  },
+  {
+    title: "The Adventures of Tom Sawyer",
+    author: "Mark Twain",
+    cover: { uri: "https://via.placeholder.com/120x180.png?text=Tom+Sawyer" },
+  },
+];
+
+
+export const allGroups = [
+  { id: 1, name: "Moby Dick Readers", relatedBook: "Moby Dick" },
+  { id: 2, name: "Romantics United", relatedBook: "Frankenstein" },
+  { id: 3, name: "Shakespeare Circle", relatedBook: "Romeo and Juliet" },
+  { id: 4, name: "Jane Austen Society", relatedBook: "Pride and Prejudice" },
+  {
+    id: 5,
+    name: "Wonderland Enthusiasts",
+    relatedBook: "Alice's Adventures in Wonderland",
+  },
+];
+
+export const allBooks = [...libraryBooks, ...recommendations];
