+++ data/libraryBooks.json
@@ -1,0 +1,37 @@
+[
+  {
+    "title": "Moby Dick",
+    "author": "Herman Melville",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Moby+Dick"
+    }
+  },
+  {
+    "title": "Frankenstein",
+    "author": "Mary Shelley",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Frankenstein"
+    }
+  },
+  {
+    "title": "Romeo and Juliet",
+    "author": "William Shakespeare",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Romeo+and+Juliet"
+    }
+  },
+  {
+    "title": "Pride and Prejudice",
+    "author": "Jane Austen",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Pride+and+Prejudice"
+    }
+  },
+  {
+    "title": "Alice's Adventures in Wonderland",
+    "author": "Lewis Carroll",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Alice"
+    }
+  }
+]
