+++ data/reccomendations.json
@@ -1,0 +1,114 @@
+[
+  {
+    "title": "Little Women",
+    "author": "Louisa May Alcott",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Little+Women"
+    }
+  },
+  {
+    "title": "Dracula",
+    "author": "Bram Stoker",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Dracula"
+    }
+  },
+  {
+    "title": "The Strange Case of Dr. Jekyll and Mr. Hyde",
+    "author": "Robert Louis Stevenson",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Jekyll+%26+Hyde"
+    }
+  },
+  {
+    "title": "The Great Gatsby",
+    "author": "F. Scott Fitzgerald",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Gatsby"
+    }
+  },
+  {
+    "title": "A Tale of Two Cities",
+    "author": "Charles Dickens",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Tale+of+Two+Cities"
+    }
+  },
+  {
+    "title": "Adventures of Huckleberry Finn",
+    "author": "Mark Twain",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Huck+Finn"
+    }
+  },
+  {
+    "title": "Wuthering Heights",
+    "author": "Emily Brontë",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Wuthering+Heights"
+    }
+  },
+  {
+    "title": "War and Peace",
+    "author": "Leo Tolstoy",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=War+and+Peace"
+    }
+  },
+  {
+    "title": "Grimms' Fairy Tales",
+    "author": "Jacob and Wilhelm Grimm",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Grimms+Tales"
+    }
+  },
+  {
+    "title": "The Adventures of Sherlock Holmes",
+    "author": "Arthur Conan Doyle",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Sherlock+Holmes"
+    }
+  },
+  {
+    "title": "The Works of Edgar Allan Poe — Volume 2",
+    "author": "Edgar Allan Poe",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Edgar+Allan+Poe+V2"
+    }
+  },
+  {
+    "title": "The Iliad",
+    "author": "Homer",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=The+Iliad"
+    }
+  },
+  {
+    "title": "The Importance of Being Earnest",
+    "author": "Oscar Wilde",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Earnest"
+    }
+  },
+  {
+    "title": "The Count of Monte Cristo",
+    "author": "Alexandre Dumas",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Monte+Cristo"
+    }
+  },
+  {
+    "title": "Anne of Green Gables",
+    "author": "L. M. Montgomery",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Anne+of+Green+Gables"
+    }
+  },
+  {
+    "title": "The Adventures of Tom Sawyer",
+    "author": "Mark Twain",
+    "cover": {
+      "uri": "https://via.placeholder.com/120x180.png?text=Tom+Sawyer"
+    }
+  }
+]
